+++ src/main/java/com/sparta/fifteen/entity/NewsFeed.java
@@ -15,14 +15,14 @@ public class NewsFeed extends Timestamped{
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     @Column(nullable = false)
-    private int authorId;
+    private long authorId;
     @Column(nullable = false)
     private String content;
     public NewsFeed(NewsFeedRequestDto newsFeedRequestDto){
         this.authorId = newsFeedRequestDto.getAuthorId();
         this.content = newsFeedRequestDto.getContent();
     }
-    public NewsFeed(int authorId, String content) {
+    public NewsFeed(long authorId, String content) {
         this.authorId = authorId;
         this.content = content;
     }
